+++ CHANGELOG.md
@@ -6,9 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [4.5.0] - 2020-10-18
+## [4.5.0] - 2020-10-21
 ### Changed
 - Enforce interception for fetch and axios for easier use - issue #19
+- Renames `websiteRootDomain` to `sessionScope`
+- Removes `refreshTokenUrl` from input and replaces it with `apiDomain` and `apiBasePath`.
+- The refresh API will alway be `apiDomain + apiBasePath + "/session/refresh"`
+- Normalizing of user input
+- Updates supported FDI to be `1.3`
+- Changes to tests to use the new config
 
 ## [4.4.1] - 2020-10-03
 ### Changed